+++ Changelog.md
@@ -4,6 +4,12 @@
 
 -   Update nrfutil device to v2.1.1.
 
+### Fixed
+
+-   Programming Modem Firmware to thingy:91 no longer shows invalid warning when
+    modem firmware is the expected one
+-   Show the device family in the Memory layout title for MCU Boot devices
+
 ## 4.2.0 - 2024-01-04
 
 ## Added